+++ monster.hpp
@@ -18,10 +18,10 @@ class Monster
     sf::Uint8 frames;
 
 public:
-    void setTexture(const sf::Texture& texture, int width, int height);
-    void set_dir(int dir);
-    void move(int x, int y);
-    void set_position(int x, int y);
+    void setTexture(const sf::Texture& texture, unsigned long width, unsigned long height);
+    void set_dir(unsigned long dir);
+    void move(unsigned long x, unsigned long y);
+    void set_position(unsigned long x, unsigned long y);
 
     bool contains(sf::Vector2f coords);
     void draw(sf::RenderWindow& window);
